part2/requirements.txt
sentence-transformers
torch
faiss-cpu
numpy
pandas 